Refactor actions folder
