+++ lib/fastlane/plugin/wpmreleasetoolkit/actions/android/an_validate_lib_strings_action.rb
@@ -1,5 +1,5 @@
 require 'fastlane/action'
-require_relative '../helper/android_localize_helper'
+require_relative '../../helper/android_localize_helper'
 
 module Fastlane
   module Actions
