+++ lib/fastlane/plugin/wpmreleasetoolkit/actions/common/create_new_milestone_action.rb
@@ -1,8 +1,8 @@
 require 'fastlane/action'
 require 'date'
-require_relative '../helper/ghhelper_helper'
-require_relative '../helper/ios/ios_version_helper'
-require_relative '../helper/android/android_version_helper'
+require_relative '../../helper/ghhelper_helper'
+require_relative '../../helper/ios/ios_version_helper'
+require_relative '../../helper/android/android_version_helper'
 module Fastlane
   module Actions
     class CreateNewMilestoneAction < Action
